Prepare for PyPI
I followed this guide as a checklist to make sure I got everything:

https://hynek.me/articles/sharing-your-labor-of-love-pypi-quick-and-dirty/

The README.rst needed a couple modifications to look right on PyPI.

1. The short description in the title in the README.
   The first title is automatically removed from the text on PyPI, but
   the text immediate after is not. Since the short description is used
   as a tagline on PyPI and was also the text immediately after the
   title, it ended up on the page twice, one after the other. I followed
   the example of Kenneth Reitz in Requests, and just put the short
   description in the title, and made a longer description.
2. PyPI doesn't like yml code blocks.
   If PyPI doesn't recognize everything in the long description as valid
   rST, it will revert it to plain text, which looks abysmal if you
   expected it to be rST. Better to have an unstyled code block than
   have it be treated as plain text.

I also made the simple case simpler in the usage instructions, and
included an advanced configuration section to showcase the usage of the
``[tox:travis]`` section in the ``tox.ini`` file.

I made it build a universal wheel, and cleaned up the description in
various places to have it only where it's needed.