Match environments based on factors
It would be very convenient if when a particular version of python is
determined from Travis, that all of the tests that contain that python
version should run. For example, if the envlist looks like this:

    py27-{django,flask,pyramid}

Then all of those three virtual environments should be run when the
Python version for Travis is Python 2.7. As it is now, the ``py27``
environment, which isn't even specified in the list, is run instead.

Taking things a bit further, if the version settings are overridden in
the ``tox:travis`` section of ``tox.ini``, then this now looks at the
envlist that is specified there, and considers its factors when
determining a matching environment. For each env listed in the
overrides, all the factors must match, but other factors may exist and
the order doesn't matter. Additionally, _any_ env in the override list
may match in order to have the env be found and run.