Drop support for Python 3.2
The latest version of pip, version 8, does not support Python 3.2 any
longer. This has manifested itself in failing builds on Travis CI. Since
pip has decided it's reasonable to drop support for Python 3.2, it
probably is for tox-travis as well.

PyPy3 still only support Python 3.2, so it is also dropped for now.