Support PyPy3 v5.2
PyPy has started releasing alpha releases of PyPy3 v5.2,
and Travis now supports `python: pypy3.3-5.2-alpha1`.

PyPy3 v5.2 depends on virtualenv>=15.0.2, whereas Travis installs 12.0.6.
Only force the use of virtualenv>=15.0.2 on PyPy3 v5.2, so that existing
users of tox-travis are not impacted.
This is complicated by Travis providing setuptools 12.0.5,
which does not support `platform_python_implementation=="PyPy"`
as an environment marker, and fails to install tox-travis
from a source tarball.  However tox-travis provides a universal wheel,
which supports that environment marker, and Travis pip 6.0.7 can install it.

This change breaks any use of tox-travis with old setuptools.
As tox-travis uses tox to install and test the package,
either we use `skipdist=True` or tell tox to use a wheel
to install tox-travis, thereby avoiding setuptools in tox.

Also, fiddle with tox internals so that it attempts to
run `python` for the Travis versions `pypy3.3-5.2-*`.