Add tox envs for docs and desc
The docs env builds the docs with sphinx. The desc env builds ensures
that the long description that will be given to PyPI is valid
reStructuredText. That would have avoided my issues today, where I had
simply forgotten to lengthen the title of one of the sections in the
HISTORY file.