Use coverage and codecov
Installing coverage_pth allows us to run scripts, including pytest and
even subprocesses, without wrapping them in the coverage program.

Convert configure into a context manager, so that it can change the
working directory back during cleanup. Lots of indentation and
formatting changes due to this requirement.

Inside the tmpdirs, create a .coveragerc file, and move them to the main
directory when cleaning up the context manager.

COVERAGE_PROCESS_START is required to be set in order to have coverage
automatically be recorded. It should be set to the configuration file to
use. We set this to .coveragerc, which also works as the default to fall
back to tox.ini and setup.cfg, though since we are writing a .coveragerc
anyway, that doesn't make a big difference to us. However, it does allow
us to not bother wrapping our invocation of pytest, since it'll read the
configuration from the tox.ini anyway.

Use parallel mode so that each subprocess gets it's own file, and they
don't clobber each other. It also makes it possible to normalize the
report output. Use coverage combine to normalize the file paths across
Tox envs, and to combine the per-process coverage files generated.

It's also good to note that the source is the named installed package.
It's sometimes suggested to use the paths to the filenames that we want
to get coverage for, but since we install the application to test, the
files in the repository aren't actually the files being run, so we need
to identify the package another way. Namely, just using the name of the
importable package.

The coverage:paths section is what reconciles the path names. Basically,
any files in a site-packages directory, that match the source (the name
of the installed package), get treated as if they were actually the
matching file in the src directory.