Deploy only a wheel to PyPI
Don't gitignore the build and dist directories, so that Travis will
clean them before it works on the deployment.

Don't install an automatically generated egg while creating the wheel.

Only upload wheels to PyPI, and since we're doing that, only run the
tests once, with the wheel, and skip the test run with the sdist
(the Tox default).