Deprecate After All
Deprecate the After All feature in favor of Travis native build stages.
Because they are native to Travis, stages are a better way to implement
waiting for all the jobs to finish.

Add a deprecation warning on usage, in the changelog, and in the docs.

Use build stages in our own deployment process, and use the new
conditional builds, stages, and jobs to limit when the deployment
happens, and which branches are built, and to better signal that
only tags should be deployed.